aws-cdk-core/pyproject.toml
[tool.poetry]
name = "preloop-aws-cdk-core"


version = "0.3.91"



description = "CDK code to create our core infra on AWS. "
authors = ["Tejas Hosangadi <tejas@preloop.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.1"
constructs = "^10"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
pip = "^23.3.2"
install = "^1.3.5"

[tool.poetry.group.dev]
optional=true

[tool.poetry.group.dev.dependencies]
pylint = "2.15.9"
numpy = "^1.22.3"
black = "22.12.0"
isort = "^5.11.4"


[tool.poetry.group.test]
optional=true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
